Renamed event test files to match project naming convention.
