+++ src/experimental/event/test/interactive/crosshair-canvas.js
