+++ src/experimental/event/test/interactive/index.js
@@ -1,3 +1,3 @@
-import renderTestSelectionPage from './renderTestSelectionPage';
+import renderTestSelectionPage from './render-test-selection-page';
 
 renderTestSelectionPage();
