+++ src/experimental/event/test/interactive/prepare-test-area.js
