+++ src/experimental/event/test/interactive/test-pages/index.js
@@ -1,0 +1,25 @@
+import legacyMouseClick from './legacy-mouse-click';
+import legacyMouseClickNoCache from './legacy-mouse-click-no-cache';
+import newMouseClick from './new-mouse-click';
+import newMouseClickAbsolute from './new-mouse-click-absolute';
+import newMouseClickFixed from './new-mouse-click-fixed';
+import newMouseDownUp from './new-mouse-down-up';
+import newMouseOverMoveOut from './new-mouse-over-move-out';
+import newMouseEnterMoveLeave from './new-mouse-enter-move-leave';
+import newTouch from './new-touch';
+import newTouchGesture from './new-touch-gesture';
+import newTouchGestureMoreFeatures from './new-touch-gesture-more-features';
+
+export default [
+  legacyMouseClick,
+  legacyMouseClickNoCache,
+  newMouseClick,
+  newMouseClickAbsolute,
+  newMouseClickFixed,
+  newMouseDownUp,
+  newMouseOverMoveOut,
+  newMouseEnterMoveLeave,
+  newTouch,
+  newTouchGesture,
+  newTouchGestureMoreFeatures
+];
