+++ src/experimental/event/test/interactive/test-pages/legacy-mouse-click.js
@@ -1,21 +1,22 @@
 /* global document */
 
-import addEvents from '../../../add-events';
-import CrosshairCanvas from '../CrosshairCanvas';
+import {addEvents as addEventsLegacy} from '../../../../../core/event';
+import CrosshairCanvas from '../crosshair-canvas';
 
 export default {
-  name: '(new) mouse click (document scroll)',
+  name: '(legacy) mouse click (document scroll, cache enabled)',
   render(testArea) {
     const canvasElement = document.createElement('canvas');
     canvasElement.height = 2 * document.documentElement.clientHeight;
     canvasElement.width = 2 * document.documentElement.clientWidth;
     testArea.appendChild(canvasElement);
     const crosshairCanvas = new CrosshairCanvas(canvasElement);
-    addEvents(canvasElement, {
-      onClick(eventInfo) {
+    addEventsLegacy(canvasElement, {
+      onClick: (eventInfo) => {
         console.log(eventInfo);
-        crosshairCanvas.setCrosshair({key: 'click', position: eventInfo.pointerPosition});
-      }
+        crosshairCanvas.setCrosshair({key: 'click', position: {x: eventInfo.x, y: eventInfo.y}});
+      },
+      centerOrigin: false
     });
   }
 };
