+++ src/experimental/event/test/interactive/test-pages/new-mouse-enter-move-leave.js
@@ -1,10 +1,10 @@
 /* global document */
 
 import addEvents from '../../../add-events';
-import CrosshairCanvas from '../CrosshairCanvas';
+import CrosshairCanvas from '../crosshair-canvas';
 
 export default {
-  name: '(new) mouse over/move/out',
+  name: '(new) mouse enter/move/leave',
   render(testArea) {
     const containerElement = document.createElement('div');
     const canvasElement = document.createElement('canvas');
@@ -25,11 +25,11 @@ export default {
     `);
     const crosshairCanvas = new CrosshairCanvas(canvasElement);
     addEvents(canvasElement, {
-      onMouseOver(eventInfo) {
+      onMouseEnter(eventInfo) {
         console.log(eventInfo);
         crosshairCanvas.reset();
         crosshairCanvas.setCrosshair({
-          key: 'over', position: eventInfo.pointerPosition, color: 'blue'
+          key: 'enter', position: eventInfo.pointerPosition, color: 'blue'
         });
       },
       onMouseMove(eventInfo) {
@@ -38,10 +38,10 @@ export default {
           key: 'move', position: eventInfo.pointerPosition, color: 'black'
         });
       },
-      onMouseOut(eventInfo) {
+      onMouseLeave(eventInfo) {
         console.log(eventInfo);
         crosshairCanvas.setCrosshair({
-          key: 'out', position: eventInfo.pointerPosition, color: 'red'
+          key: 'leave', position: eventInfo.pointerPosition, color: 'red'
         });
       }
     });
